automod: move content fetching to new file
